Small modifications regarding cost calculations.
Floating types for execution contribution and weights of edges
Some name modifications with better semantics

+++ application.h
@@ -15,7 +15,7 @@ class Application
 		Application(){}
 		Application(unsigned int nodes);
 		unsigned int getTotalFunctions(){ return _TotalFunctions; };
-		unsigned int getFunctionContrib(unsigned int fno){ return _Functions[fno].getExecContrib(); }
+		float getFunctionContrib(unsigned int fno){ return _Functions[fno].getExecContrib(); }
 		unsigned long long getEdgeWeight(unsigned int i, unsigned int j){return _Edges[i][j].getWeight();}
 		void print();
 };
